requirements.txt
Appium-Python-Client==1.0.2
attrs==23.2.0
certifi==2024.2.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==0.13.1
py==1.11.0
PySocks==1.7.1
pytest==6.1.1
selenium==3.14.1
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==1.26.18
wsproto==1.2.0
